Style Font for Header/Hero and Arrange Layout
Changed the class title of "information" to "hero" as it should be.
Added many classes to control specific parts of text, although
this may end up being a clunky method. Set the font styles for color
and size on most of the Header and Hero elements.

Also added left and right margins to the entire body, min-height set to
99% vh, and justify-content to space-between to make the overall layout
of the page reflect what the final product will look like.