Set Up Information Container
Lot of changes in this one. I added placeholder images, and changed
the html to locate them properly.

Set up the information section using a column flex container, and
jankily adjusted the sizing and max width of the child containers so
that the subtext under each illustration lined up properly. It's kind of
a mess but it works. Also added borders to the placeholder images.

Definitely not the most efficient way to do all this but it is working.
Need to go back and remove some redundant code that probably isn't doing
anything. I am just committing it how it is right now because I got it
looking how I want.