Configure Quote Section
Created a class for the quote text and configured the font properties
to make it look how I want.

Adding padding to the quote container with flex displayed, wrap on,
and justify-content set to flex-end so I can get the attribution to
line up just how I want underneath the quote text.

A problem I am having with this section, however, is this font family
does not seem to be able to display a lighter font-weight so it does
not look *exactly* how I want it to.