Style Call to Action Section
Flexed the parent container for this section with a blue background,
and a margin to add white space surrounding it. Also added a border
radius. Flexed the interior text container with a column layout.
Styled the sign up button with blue background and border radius.

Couldn't get the spacing exactly how I wanted although I think I could
with enough time. I like how it looks though so just moving on to the
next section.